+++ backend/src/modules/expense/expense.contoller.ts
@@ -53,7 +53,11 @@ const ExpenseController = {
 
   handleGetById: async (req: CustomRequest, res: Response) => {
     const { id } = req.params;
-    const expense = await ExpenseService.getById(Number(id));
+    const organizationId = req.user?.organizationId;
+
+    if (!organizationId)
+      throw new BadRequestError("Organization Id is not found");
+    const expense = await ExpenseService.getById(Number(id), organizationId);
     res
       .status(200)
       .json({ message: "Expense fetched Successfully", data: expense });