+++ backend/src/modules/expense/expense.repository.ts
@@ -53,6 +53,7 @@ const ExpenseRepository = {
         remarks: existingData.remarks || undefined,
         subCommitteeId: existingData.subCommitteeId || undefined,
         payerId: existingData.payerId || undefined,
+        
       };
 
       // console.log({ oldData });