+++ backend/src/modules/expense/expense.schema.ts
@@ -3,8 +3,8 @@ import z from "zod";
 const expenseSchema = z.object({
   date: z.iso.datetime({ error: "Invalid date format" }),
   nepaliDate: z.string().min(1, "Date is required"),
-  name: z.string().min(1, "Name is required"),
-  address: z.string().min(1, "Address is required"),
+  recepientName: z.string().min(1, "Name is required"),
+  recepientAddress: z.string().min(1, "Address is required"),
   particulars: z.string().min(1, "Particulars is required"),
   quantity: z.string().optional(),
   amount: z