+++ frontend/src/components/common/nepali-date-input.tsx
@@ -25,11 +25,11 @@ const NepaliDateInput = ({
   
   useEffect(() => {
     onValueChange(inputDate)
-  }, [inputDate])
+  }, [inputDate, onValueChange])
 
   const handleDateChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    let value = event.currentTarget.value
-    let keyPressed = (event.nativeEvent as InputEvent).inputType
+    const value = event.currentTarget.value
+    const keyPressed = (event.nativeEvent as InputEvent).inputType
 
     setInputDate((prevValue) => {
       if (