+++ frontend/src/components/layouts/site-header.tsx
@@ -14,7 +14,7 @@ export function SiteHeader() {
 
   useEffect(() => {
     if (role !== "OWNER") dispatch(fetchUserAssignedOrganization())
-  }, [isUserAssignedOrgFetchIdle, dispatch])
+  }, [isUserAssignedOrgFetchIdle, dispatch, role])
   return (
     <header className="flex h-(--header-height) shrink-0 items-center gap-2 border-b transition-[width,height] ease-linear group-has-data-[collapsible=icon]/sidebar-wrapper:h-(--header-height)">
       <div className="flex w-full items-center gap-1 px-4 lg:gap-2 lg:px-6">