+++ frontend/src/main.tsx
@@ -5,7 +5,6 @@ import { Provider } from "react-redux"
 import router from "./routes/routes.tsx"
 import { createRoot } from "react-dom/client"
 import { RouterProvider } from "react-router-dom"
-import { AuthProvider } from "./context/authContext.tsx"
 import { TooltipProvider } from "./components/ui/tooltip.tsx"
 import { ThemeProvider } from "@/components/common/theme-provider.tsx"
 import {
@@ -56,13 +55,11 @@ createRoot(document.getElementById("root")!).render(
   <StrictMode>
     <Provider store={store}>
       <QueryClientProvider client={queryClient}>
-        <AuthProvider>
           <ThemeProvider>
             <TooltipProvider>
               <RouterProvider router={router} />
             </TooltipProvider>
           </ThemeProvider>
-        </AuthProvider>
       </QueryClientProvider>
     </Provider>
   </StrictMode>