+++ frontend/src/page/category/useCreateCategoryForm.tsx
@@ -1,10 +1,9 @@
-import { useForm, type SubmitHandler } from "react-hook-form"
+import { useForm } from "react-hook-form"
 import { useMutation, useQueryClient } from "@tanstack/react-query"
 import { zodResolver } from "@hookform/resolvers/zod"
 import CategorySchema from "./category.schema"
 import CategoryRepository from "./category.repository"
 import { toast } from "sonner"
-import type { TCreateCategory } from "./category.types"
 
 const useCreateCategory = () => {
   const queryClient = useQueryClient()
@@ -18,13 +17,11 @@ const useCreateCategory = () => {
   })
   const form = useForm({ resolver: zodResolver(CategorySchema) })
 
-  const onSubmit: SubmitHandler<TCreateCategory> = (data) => {
-    mutate(data)
-  }
+
 
   return {
     ...form,
-    onSubmit,
+    mutate,
     isPending,
     isSuccess,
     isError,